Added tests for the LatLng. Changed supported version of Flow.
